requirements.txt
flask
flask_sqlalchemy
pandas
openpyxl
python-telegram-bot==13.15